% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionDataMultiOutput.R
\name{as.PredictionDataMultioutput}
\alias{as.PredictionDataMultioutput}
\title{Convert a data.table prediction object from learners that inherently can do e.g. multilabel.}
\usage{
as.PredictionDataMultioutput(data, target_types)
}
\arguments{
\item{data}{\code{\link{data.table}}\cr
Contains predictions, must have row_id.}

\item{task_types}{\code{\link{character}}\cr
Named character vector of per-target task-types.
E.g. c(tgt1 = "regr", tgt2 = "classif")}
}
\description{
Convert a data.table prediction object from learners that inherently can do e.g. multilabel.
}
