% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskMultiOutput.R
\name{convert_to_basic_tasks}
\alias{convert_to_basic_tasks}
\title{Convert to basic task_type}
\usage{
convert_to_basic_tasks(task)
}
\arguments{
\item{task}{\code{\link{TaskMultiOutput}}\cr
Multi-output task to split into tasks of type \code{task$task_types}.}
}
\description{
Splits a \code{\link{TaskMultiOutput}} into several \code{\link{Task}}s of type "regr", "classif", ...
according to "task$task_types".
}
