% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3multioutput-package}
\alias{mlr3multioutput}
\alias{mlr3multioutput-package}
\title{mlr3multioutput: Multioutput extension for 'mlr3'}
\description{
Extends the 'mlr3' package with multi-output tasks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https:///mlr3multioutput.mlr-org.com}
  \item \url{https://github.com/mlr-org/multioutput}
  \item Report bugs at \url{https://github.com/mlr-org/multioutput/issues}
}

}
\author{
\strong{Maintainer}: Michel Lang \email{michellang@gmail.com} (\href{https://orcid.org/0000-0001-9754-0393}{ORCID})

Authors:
\itemize{
  \item Florian Pfisterer \email{pfistererf@googlemail.com}
}

}
