% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskMultiOutput_flags.R
\name{mlr_tasks_flags}
\alias{mlr_tasks_flags}
\title{flags Task}
\format{
\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskMultiOutput}.
}
\description{
A multi-output task for the \link[mldr.datasets:flags]{mldr.datasets::flags} data set.
}
\section{Construction}{
\preformatted{mlr_tasks$get("flags")
tsk("flags")
}
}

