% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskMultiOutput_linnerud.R
\name{mlr_tasks_linnerud}
\alias{mlr_tasks_linnerud}
\title{Linnerud Task}
\format{
\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskMultiOutput}.
}
\description{
A multi-output task for the \link[plsdepot:linnerud]{plsdepot::linnerud} data set.
}
\section{Construction}{
\preformatted{mlr_tasks$get("linnerud")
tsk("linnerud")
}
}

