% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionMultiLabel.R
\name{PredictionClassif}
\alias{PredictionClassif}
\title{Prediction Object for Multilabel Analysis}
\description{
This object wraps the predictions returned by a learner of class \link{LearnerMultiOutput}, i.e.
the predicted partition and MultiOutputer probability.
}
\seealso{
Other Prediction: 
\code{\link{PredictionMultiOutput}}
}
\concept{Prediction}
\section{Super class}{
\code{\link[mlr3:Prediction]{mlr3::Prediction}} -> \code{PredictionMultiLabel}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{response}}{(\code{factor()})\cr
Access to the stored predicted class labels.}

\item{\code{prob}}{(\code{matrix()})\cr
Access to the stored probabilities.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PredictionClassif$new()}}
\item \href{#method-set_threshold}{\code{PredictionClassif$set_threshold()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Prediction" data-id="format">}\href{../../mlr3/html/Prediction.html#method-format}{\code{mlr3::Prediction$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Prediction" data-id="help">}\href{../../mlr3/html/Prediction.html#method-help}{\code{mlr3::Prediction$help()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Prediction" data-id="print">}\href{../../mlr3/html/Prediction.html#method-print}{\code{mlr3::Prediction$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Prediction" data-id="score">}\href{../../mlr3/html/Prediction.html#method-score}{\code{mlr3::Prediction$score()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionClassif$new(
  task = NULL,
  row_ids = task$row_ids,
  truth = task$truth(),
  response = NULL,
  prob = NULL,
  check = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{(\link{TaskClassif})\cr
Task, used to extract defaults for \code{row_ids} and \code{truth}.}

\item{\code{row_ids}}{(\code{integer()})\cr
Row ids of the predicted observations, i.e. the row ids of the test set.}

\item{\code{truth}}{(\code{factor()})\cr
Matrix of True (observed) labels. See the note on manual construction.}

\item{\code{response}}{(\code{character()} | \code{factor()})\cr
Matrix of predicted class labels.
One vector for each label in the test set.
Character vectors are automatically converted to factors.
See the note on manual construction.}

\item{\code{prob}}{(\code{matrix()})\cr
Numeric matrix of posterior class probabilities with one column for each class
and one row for each observation in the test set.
Columns must be named with class labels, row names are automatically removed.
If \code{prob} is provided, but \code{response} is not, the class labels are calculated from
the probabilities using \code{\link[=max.col]{max.col()}} with \code{ties.method} set to \code{"random"}.}

\item{\code{check}}{(\code{logical(1)})\cr
If \code{TRUE}, performs some argument checks and predict type conversions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_threshold"></a>}}
\if{latex}{\out{\hypertarget{method-set_threshold}{}}}
\subsection{Method \code{set_threshold()}}{
Sets the prediction response based on the provided threshold.
See the section on thresholding for more information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionClassif$set_threshold(threshold, ties_method = "random")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{threshold}}{(\code{numeric()}).}

\item{\code{ties_method}}{(\code{character(1)})\cr
One of \code{"random"}, \code{"first"} or \code{"last"} (c.f. \code{\link[=max.col]{max.col()}}) to determine how to deal with
tied probabilities.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns the object itself, but modified \strong{by reference}.
You need to explicitly \verb{$clone()} the object beforehand if you want to keeps
the object in its previous state.
}
}
}
