% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionMultioutput.R
\name{PredictionMultioutput}
\alias{PredictionMultioutput}
\title{Prediction Object for MultiOutputer Analysis}
\description{
This object wraps the predictions returned by a learner of class \link{LearnerMultiOutput}, i.e.
the predicted partition and MultiOutputer probability.
}
\concept{Prediction}
\section{Super class}{
\code{\link[mlr3:Prediction]{mlr3::Prediction}} -> \code{PredictionMultioutput}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{predictions}}{(\code{list()})\cr
Access the stored predictions.}

\item{\code{missing}}{(\code{integer()})\cr
Returns \code{row_ids} for which the predictions are missing or incomplete.}

\item{\code{row_ids}}{(\code{integer()})\cr
Access the stored row_ids.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PredictionMultioutput$new()}}
\item \href{#method-print}{\code{PredictionMultioutput$print()}}
\item \href{#method-score_separate}{\code{PredictionMultioutput$score_separate()}}
\item \href{#method-clone}{\code{PredictionMultioutput$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Prediction" data-id="format">}\href{../../mlr3/html/Prediction.html#method-format}{\code{mlr3::Prediction$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Prediction" data-id="help">}\href{../../mlr3/html/Prediction.html#method-help}{\code{mlr3::Prediction$help()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Prediction" data-id="score">}\href{../../mlr3/html/Prediction.html#method-score}{\code{mlr3::Prediction$score()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionMultioutput$new(
  task = NULL,
  row_ids = task$row_ids,
  predictions = list(),
  check = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{(\link{TaskMultioutput})\cr
Task, used to extract defaults for \code{row_ids}.}

\item{\code{row_ids}}{(\code{integer()})\cr
Row ids of the predicted observations, i.e. the row ids of the test set.}

\item{\code{predictions}}{(\code{list()})\cr
(Named) list of per-target predictions. Used to construct the \code{Prediction}-object.}

\item{\code{check}}{(\code{logical(1)})\cr
If \code{TRUE}, performs argument checks and predict type conversions.}

\item{\code{...}}{(\code{list()})\cr
(Named) list of per-target truths. Only used for compatibility with \code{Prediction$new()}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Printer for the Prediction object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionMultioutput$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(\code{any})\cr
Not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-score_separate"></a>}}
\if{latex}{\out{\hypertarget{method-score_separate}{}}}
\subsection{Method \code{score_separate()}}{
Returns scores for each measure separately.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionMultioutput$score_separate(measures, task)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{measures}}{\code{list}\cr
List of \code{\link{MeasureMultioutput}} to score.}

\item{\code{task}}{\code{\link{TaskMultioutput}}\cr
Task to use for scoring}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{numeric()} vector of scores.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionMultioutput$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
