% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureMultiOutputWeightedAvg.R
\name{MeasureMultiOutputWeightedAvg}
\alias{MeasureMultiOutputWeightedAvg}
\title{MultiOutput weighted average Measure}
\description{
Computes a weighted average over measured scores for each target.
\itemize{
\item measures that should be maximized are automatically multiplied by '-1' internally during aggregation
and \code{minimize} is therefore set to \code{TRUE}
\item \code{task_type} is set to \code{"multioutput"}.
\item Possible values for \code{predict_type} are all values from \code{mlr_reflections$learner_predict_types}.
They are currently collected by accessing each \code{\link{Measure}}s "predict_type" slot.
Currently limited to only a single 'predict_type' across all measures.
\item \code{packages} are all packages required from the supplied measures.
\item The \code{range} is automatically computed based on the specified \code{weights}.
}

Predefined measures can be found in the \link[mlr3misc:Dictionary]{mlr3misc::Dictionary} \link[mlr3:mlr_measures]{mlr3::mlr_measures}.
}
\seealso{
Example measures:
}
\section{Super classes}{
\code{\link[mlr3:Measure]{mlr3::Measure}} -> \code{\link[mlr3multioutput:MeasureMultiOutput]{mlr3multioutput::MeasureMultiOutput}} -> \code{MeasureMultiOutputWeightedAvg}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{measures}}{(\code{list()})\cr
Access the stored measures.}

\item{\code{weights}}{(\code{numeric()})\cr
Access the stored weights.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MeasureMultiOutputWeightedAvg$new()}}
\item \href{#method-score_separate}{\code{MeasureMultiOutputWeightedAvg$score_separate()}}
\item \href{#method-clone}{\code{MeasureMultiOutputWeightedAvg$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="aggregate">}\href{../../mlr3/html/Measure.html#method-aggregate}{\code{mlr3::Measure$aggregate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="format">}\href{../../mlr3/html/Measure.html#method-format}{\code{mlr3::Measure$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="help">}\href{../../mlr3/html/Measure.html#method-help}{\code{mlr3::Measure$help()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="print">}\href{../../mlr3/html/Measure.html#method-print}{\code{mlr3::Measure$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="score">}\href{../../mlr3/html/Measure.html#method-score}{\code{mlr3::Measure$score()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureMultiOutputWeightedAvg$new(
  name = "weightedavg",
  measures,
  weights = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{(\code{character(1)})\cr
Name of the measure. Default: "weightedavg".}

\item{\code{measures}}{(\code{\link{list}})\cr
Named list of measures to be applied to each "target".\cr
Either named with \code{target_names}, mapping targets to measures or
named with \code{task_types}, defining one measure per task_type.
Defaults to \code{mlr_reflections$default_measures} for each task type.}

\item{\code{weights}}{(\code{numeric()})\cr
Named numeric defining weights for respective 'measures'.\cr
Either named with \code{target_names}, mapping targets to weights or
named with \code{task_types}, defining one weight per task_type.
Defaults to \code{1} for each measure.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{\link{MeasureMultiOutput}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-score_separate"></a>}}
\if{latex}{\out{\hypertarget{method-score_separate}{}}}
\subsection{Method \code{score_separate()}}{
Returns scores for each measure in self$measures separately.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureMultiOutputWeightedAvg$score_separate(prediction, task, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prediction}}{\code{\link{PredictionMultiOutput}}\cr
Prediction to score.}

\item{\code{task}}{\code{\link{TaskMultiOutput}}\cr
Task to score.}

\item{\code{...}}{(\code{any})\cr
Currently not used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{numeric()} vector of scores.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureMultiOutputWeightedAvg$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
